Reworking backend for sprint 6

+++ backend/src/main/java/de/fraunhofer/iao/querimonia/db/repositories/ActionRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.repository.CrudRepository;
 /**
  * Interface to access the action table.
  */
-public interface ActionRepository extends CrudRepository<Action, Integer> {
+public interface ActionRepository extends CrudRepository<Action, Long> {
 }
