+++ backend/src/main/java/de/fraunhofer/iao/querimonia/db/repositories/ConfigurationRepository.java
@@ -3,5 +3,5 @@ package de.fraunhofer.iao.querimonia.db.repositories;
 import de.fraunhofer.iao.querimonia.config.Configuration;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ConfigurationRepository extends CrudRepository<Configuration, Integer> {
+public interface ConfigurationRepository extends CrudRepository<Configuration, Long> {
 }
